+++ mvc/src/main/java/nextstep/mvc/controller/asis/ControllerHandlerAdapter.java
@@ -1,13 +1,12 @@
-package com.techcourse;
+package nextstep.mvc.controller.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class ManualHandlerAdapter implements HandlerAdapter {
+public class ControllerHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
